requirements.txt
drf-yasg==1.21.7
inflection==0.5.1
packaging==21.3
pyparsing==3.0.7
pyyaml==6.0.1
uritemplate==4.1.1
Django~=3.2.23
djangorestframework~=3.14.0